+++ lib/rspec/active_record/create_record.rb
@@ -3,18 +3,35 @@
 module RSpec
   module ActiveRecord
     # A matcher
-    class CreateRecord
-      include RSpec::Matchers::Composable
+    class CreateRecord < Matcher
       include RSpec::Matchers::BuiltIn::BaseMatcher::HashFormatting
 
       def initialize(scope)
         @scope = scope
+        super()
       end
 
       # Make sure that created record matches attributes
       def matching(attributes)
         @attributes = attributes
         self
+      end
+
+      def times(times)
+        @times = times
+        self
+      end
+
+      def once
+        times(1)
+      end
+
+      def twice
+        times(2)
+      end
+
+      def thrice
+        times(3)
       end
 
       def supports_block_expectations?
@@ -25,25 +42,25 @@ module RSpec
         existing_ids = @scope.ids
         block.call
 
-        @new_records = @scope.where.not(@scope.primary_key => existing_ids)
+        @new_records = @scope.where.not(@scope.primary_key => existing_ids).to_a
 
-        if @attributes
-          match_attributes?
-        else
-          @new_records.present?
-        end
+        match_times? && match_attributes? && @new_records.present?
       end
 
       def description
         message = "create #{scope_name}"
-        if @attributes
-          message += " matching #{RSpec::Support::ObjectFormatter.format(surface_descriptions_in(@attributes))}"
-        end
+        message += " #{@times} time#{"s" if @times != 1}" if @times
+        message += " matching #{format_hash(@attributes)}" if @attributes
         improve_hash_formatting message
       end
 
       def failure_message
-        add_diff "expected to #{description} but did not"
+        failure = "expected to #{description} but"
+        if !match_times?
+          "#{failure} created #{@new_records.size}"
+        else
+          add_diff "#{failure} did not"
+        end
       end
 
       def failure_message_when_negated
@@ -57,8 +74,13 @@ module RSpec
         message
       end
 
+      def match_times?
+        @times.nil? || @times == @new_records.size
+      end
+
       def match_attributes?
-        @new_records.any? { |record| RSpec::Matchers::BuiltIn::HaveAttributes.new(@attributes).matches?(record) }
+        @attributes.nil? ||
+          @new_records.any? { |record| RSpec::Matchers::BuiltIn::HaveAttributes.new(@attributes).matches?(record) }
       end
 
       def scope_name